chore: clean up index page code
